# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Startup Review Guy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Founder {
  id: string;
  startupId: string;
  name: string;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  photoUrl?: string | null;
}

export interface FounderInput {
  name: string;
  role?: string;
  linkedinUrl?: string;
  photoUrl?: string;
}

export interface StartupSummary {
  id: string;
  slug: string;
  name: string;
  tagline: string;
  category: string;
  tags?: string[];
  /** @nullable */
  logoUrl?: string | null;
  youtubeVideoId: string;
  /** @nullable */
  fundingStage?: string | null;
  reviewedAt: string;
  published: boolean;
}

export interface Startup {
  id: string;
  slug: string;
  name: string;
  tagline: string;
  description: string;
  /** @nullable */
  logoUrl?: string | null;
  youtubeVideoId: string;
  category: string;
  tags?: string[];
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  twitterUrl?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  founders?: Founder[];
  /** @nullable */
  fundingStage?: string | null;
  /** @nullable */
  totalRaised?: string | null;
  /** @nullable */
  notableInvestors?: string | null;
  /** @nullable */
  lastRoundDate?: string | null;
  /** @nullable */
  revenueArr?: string | null;
  revenueEstimated?: boolean;
  /** @nullable */
  competitors?: string | null;
  /** @nullable */
  useCase?: string | null;
  /** @nullable */
  verdict?: string | null;
  reviewedAt: string;
  updatedAt: string;
  published: boolean;
  relatedStartups?: StartupSummary[];
}

export interface StartupListResponse {
  startups: StartupSummary[];
  total: number;
  page: number;
  limit: number;
}

export interface StartupInput {
  slug: string;
  name: string;
  tagline: string;
  description: string;
  logoUrl?: string;
  youtubeVideoId: string;
  category: string;
  tags?: string[];
  websiteUrl?: string;
  twitterUrl?: string;
  linkedinUrl?: string;
  founders?: FounderInput[];
  fundingStage?: string;
  totalRaised?: string;
  notableInvestors?: string;
  lastRoundDate?: string;
  revenueArr?: string;
  revenueEstimated?: boolean;
  competitors?: string;
  useCase?: string;
  verdict?: string;
  published?: boolean;
}

export interface StartupPatch {
  slug?: string;
  name?: string;
  tagline?: string;
  description?: string;
  logoUrl?: string;
  youtubeVideoId?: string;
  category?: string;
  tags?: string[];
  websiteUrl?: string;
  twitterUrl?: string;
  linkedinUrl?: string;
  founders?: FounderInput[];
  fundingStage?: string;
  totalRaised?: string;
  notableInvestors?: string;
  lastRoundDate?: string;
  revenueArr?: string;
  revenueEstimated?: boolean;
  competitors?: string;
  useCase?: string;
  verdict?: string;
  published?: boolean;
}

export interface PartnerInquiry {
  id: string;
  companyName: string;
  contactName: string;
  email: string;
  /** @nullable */
  website?: string | null;
  message: string;
  /** @nullable */
  budgetRange?: string | null;
  createdAt: string;
  status: string;
}

export interface PartnerInquiryInput {
  companyName: string;
  contactName: string;
  email: string;
  website?: string;
  message: string;
  budgetRange?: string;
  honeypot?: string;
}

export interface InquiryStatusUpdate {
  status: string;
}

export interface SiteStats {
  totalReviews: number;
  totalCategories: number;
  /** @nullable */
  subscriberCount?: number | null;
  /** @nullable */
  totalViews?: number | null;
}

export type AuthSessionUser = {
  email?: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  picture?: string | null;
};

export interface AuthSession {
  authenticated: boolean;
  user?: AuthSessionUser;
}

export interface LogoutResult {
  success: boolean;
}

export type ListStartupsParams = {
category?: string;
fundingStage?: string;
search?: string;
page?: number;
limit?: number;
};

export type AdminListStartupsParams = {
page?: number;
limit?: number;
search?: string;
};

export type AdminListInquiriesParams = {
status?: string;
};