# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Startup Review Guy API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthSessionUser } from './authSessionUser';

export interface AuthSession {
  authenticated: boolean;
  user?: AuthSessionUser;
}