# lib/api-zod/src/generated/types/founder.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Startup Review Guy API
 * OpenAPI spec version: 0.1.0
 */

export interface Founder {
  id: string;
  startupId: string;
  name: string;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  photoUrl?: string | null;
}