# lib/api-zod/src/generated/types/partnerInquiry.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Startup Review Guy API
 * OpenAPI spec version: 0.1.0
 */

export interface PartnerInquiry {
  id: string;
  companyName: string;
  contactName: string;
  email: string;
  /** @nullable */
  website?: string | null;
  message: string;
  /** @nullable */
  budgetRange?: string | null;
  createdAt: string;
  status: string;
}