# lib/api-zod/src/generated/types/siteStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Startup Review Guy API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteStats {
  totalReviews: number;
  totalCategories: number;
  /** @nullable */
  subscriberCount?: number | null;
  /** @nullable */
  totalViews?: number | null;
}