# lib/api-zod/src/generated/types/startup.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Startup Review Guy API
 * OpenAPI spec version: 0.1.0
 */
import type { Founder } from './founder';
import type { StartupSummary } from './startupSummary';

export interface Startup {
  id: string;
  slug: string;
  name: string;
  tagline: string;
  description: string;
  /** @nullable */
  logoUrl?: string | null;
  youtubeVideoId: string;
  category: string;
  tags?: string[];
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  twitterUrl?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  founders?: Founder[];
  /** @nullable */
  fundingStage?: string | null;
  /** @nullable */
  totalRaised?: string | null;
  /** @nullable */
  notableInvestors?: string | null;
  /** @nullable */
  lastRoundDate?: string | null;
  /** @nullable */
  revenueArr?: string | null;
  revenueEstimated?: boolean;
  /** @nullable */
  competitors?: string | null;
  /** @nullable */
  useCase?: string | null;
  /** @nullable */
  verdict?: string | null;
  reviewedAt: string;
  updatedAt: string;
  published: boolean;
  relatedStartups?: StartupSummary[];
}