# lib/api-zod/src/generated/types/startupInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Startup Review Guy API
 * OpenAPI spec version: 0.1.0
 */
import type { FounderInput } from './founderInput';

export interface StartupInput {
  slug: string;
  name: string;
  tagline: string;
  description: string;
  logoUrl?: string;
  youtubeVideoId: string;
  category: string;
  tags?: string[];
  websiteUrl?: string;
  twitterUrl?: string;
  linkedinUrl?: string;
  founders?: FounderInput[];
  fundingStage?: string;
  totalRaised?: string;
  notableInvestors?: string;
  lastRoundDate?: string;
  revenueArr?: string;
  revenueEstimated?: boolean;
  competitors?: string;
  useCase?: string;
  verdict?: string;
  published?: boolean;
}