# lib/api-zod/src/generated/types/startupSummary.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Startup Review Guy API
 * OpenAPI spec version: 0.1.0
 */

export interface StartupSummary {
  id: string;
  slug: string;
  name: string;
  tagline: string;
  category: string;
  tags?: string[];
  /** @nullable */
  logoUrl?: string | null;
  youtubeVideoId: string;
  /** @nullable */
  fundingStage?: string | null;
  reviewedAt: string;
  published: boolean;
}